chore: Add all of the converters into bot/utils/converters.py
